% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{bin_interactions}
\alias{bin_interactions}
\title{Bins putative interactions by distance}
\usage{
bin_interactions(interactions, fragments, bins = 50, log_file = NA)
}
\arguments{
\item{bins}{Number of bins to place the interactions into.}

\item{log_file}{Path to a log file.}

\item{interactions_file}{Data table containing putative interactions. Columns called baitID, preyID, N, storing the bait fragment ID, prey fragment ID and readcount, respectively.}

\item{fragments_file}{Data table containing fragment information. Columns called chrom, chromStart, chromEnd, ID, storing the chromosome, starting coordinate, ending coordinate and ID of a fragment, respectively.}
}
\value{
List containing the binned interactions ($binned_interactions) and an overview of bins ($overview).
}
\description{
Places putative interactions into equally-sized bins based on the distances they span.
}
\examples{
base = system.file("extdata",package="peaky")

interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

\donttest{BI = bin_interactions(interactions, fragments, bins=20)}
print(BI)

}
