% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{interpret_peaky}
\alias{interpret_peaky}
\title{Interprets RJMCMC results}
\usage{
interpret_peaky(rjmcmclist, index, baits_dir, output_dir, omega_power)
}
\arguments{
\item{rjmcmclist}{Path to the list of RJMCMC results to be interpreted.}

\item{index}{Which RJMCMC result on the list to process, with 1 corresponding to the first one on the list.}

\item{output_dir}{Directory where the merged RJMCMC results and bait information will be stored, one file per bait. Will be created if it does not exist.}

\item{omega_power}{The same value as used when running peaky, i.e. the expected decay of adjusted read counts around a truly interacting prey.}
}
\value{
The output directory.
}
\description{
Extracts the marginal posterior probabilities of contact and other relevant statistics from the RJMCMC results and merges these with bait-associated information.
}
\examples{
base = "/local/data/public/hic/demo"
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

\dontrun{bin_interactions(interactions_file, fragments_file, output_dir=bins_dir)}

}
