% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{interpret_peaky_fs}
\alias{interpret_peaky_fs}
\title{Wrapper for interpret_peaky() that interacts with the filesystem}
\usage{
interpret_peaky_fs(rjmcmclist, index, baits_dir, output_dir, omega_power)
}
\arguments{
\item{rjmcmclist}{Path to the list of RJMCMC results to be interpreted.}

\item{index}{Which RJMCMC result on the list to process, with 1 corresponding to the first one on the list.}

\item{output_dir}{Directory where the merged RJMCMC results and bait information will be stored, one file per bait. Will be created if it does not exist.}

\item{omega_power}{The same value as used when running peaky, i.e. the expected decay of adjusted read counts around a truly interacting prey. See details.}
}
\value{
The output directory.

Details.
The steepness of the function to be fitted to putative peaks is determined by \eqn{\omega} according to \eqn{\beta \exp{- \abs{\omega * d}}}, where \eqn{\beta} represents peak height and \eqn{d} the distance from the center of the peak in bp.
}
\description{
Extracts the marginal posterior probabilities of contact and other relevant statistics from the RJMCMC results and merges these with bait-associated information.
}
\examples{
base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

\donttest{
bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

fits_dir = paste0(base,"/fits")

for(bin_index in 1:8){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir)
}

baits_dir = paste0(base,"/baits")

split_baits_fs(bins_dir,residuals_dir = fits_dir, indices=1:8, output_dir = baits_dir)

baitlist = paste0(baits_dir,"/baitlist.txt")
rjmcmc_dir = paste0(base,"/rjmcmc")
omega_power = 4.7

for(i in 1:49){
 peaky_fs(baitlist,i,output_dir=rjmcmc_dir,omega_power=omega_power)
}

rjmcmclist = paste0(rjmcmc_dir,"/rjmcmclist.txt")
baits_rjmcmc_dir = paste0(base,"/baits_rjmcmc")
interpret_peaky(rjmcmclist,1,omega_power,baits_dir,baits_rjmcmc_dir)
}
}
