% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{model_bin_fs}
\alias{model_bin_fs}
\title{Wrapper for model_bin() that interacts with the filesystem}
\usage{
model_bin_fs(
  bins_dir,
  bin_index,
  output_dir,
  subsample_size = NA,
  gamlss_cycles = 200,
  gamlss_crit = 0.1,
  formula_add = NA,
  formula_replace = NA
)
}
\arguments{
\item{bins_dir}{Directory containing putative interactions that are binned by distance.}

\item{bin_index}{Index of the bin to be processed.}

\item{output_dir}{Directory where the parametrized model and adjusted readcounts will be stored. Will be created if it does not exist.}

\item{subsample_size}{Number of interactions based on which the null-model is parametrized. By default, all are used.}

\item{gamlss_cycles}{GAMLSS maximum number of cycles for convergence (see gamlss::gamlss.control).}

\item{gamlss_crit}{GAMLSS convergence criterion (see gamlss::gamlss.control).}

\item{formula_add}{Additional covariates for the default model formula. To add covariates A and B when modelling counts, supply "A + B".}

\item{formula_replace}{Replaces the default model formula entirely. Suppled in the form "N ~ A + B". Takes precedence over formula_add.}
}
\value{
List containing the path to the output directory, the null model, and the adjusted read counts.
}
\description{
Parametrizes a negative binomial null model for the readcounts in a given distance bin. Covariates are the distance between an interaction's bait and prey fragments, their lengths, and transchromosomal bait activity. Randomized quantile residuals computed from this model are taken as noise-adjusted readcounts and stored on disk.
}
\examples{
base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

\donttest{bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)}

fits_dir = paste0(base,"/fits")

for(bin_index in 1:8){
  \donttest{model_bin_fs(bins_dir,bin_index,output_dir=fits_dir)}
}
}
