% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{peaky}
\alias{peaky}
\title{Regroups putative interactions by bait}
\usage{
peaky(baitlist, index, output_dir, omega_power, min_interactions = 20,
  log_file = TRUE, iterations = 1e+06)
}
\arguments{
\item{baitlist}{Path to the list of baits to be processed.}

\item{index}{Which bait on the list to process, with 1 corresponding to the first one on the list.}

\item{output_dir}{Directory where all RJMCMC results will be stored. Will be created if it does not exist.}

\item{omega_power}{Expected decay of adjusted read counts around a truly interacting prey. See details.}

\item{min_interactions}{Whether adjusted readcounts are to be plotted aganst distance and stored for each bait.}

\item{log_file}{Whether or not a log-file is required.}

\item{min_interactions}{Minimum requirement for the number of prey fragments (and thus counts) associated with a bait, baits with fewer are skipped.}
}
\value{
The output directory.

Details.
The steepness of the function to be fitted to putative peaks, \eqn{\beta \exp{- \abs{\omega * d}}}.
}
\description{
Regroups putative interactions by bait, rather than by distance, to prepare them for RJMCMC processing.
}
\examples{
base = "/local/data/public/hic/demo"
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

\dontrun{bin_interactions(interactions_file, fragments_file, output_dir=bins_dir)}

}
