% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{peaky_fs}
\alias{peaky_fs}
\title{Wrapper for peaky() that interacts with the filesystem}
\usage{
peaky_fs(baitlist, index, output_dir, omega_power, iterations = 1e+06,
  min_interactions = 20)
}
\arguments{
\item{baitlist}{Path to the list of baits to be processed.}

\item{index}{Which bait on the list to process, with 1 corresponding to the first one on the list.}

\item{output_dir}{Directory where all RJMCMC results will be stored. Will be created if it does not exist.}

\item{omega_power}{Expected decay of adjusted read counts around a truly interacting prey. See details.}

\item{iterations}{Number of models to parametrize. Greated numbers should lead to increased reproducibility.}

\item{min_interactions}{Minimum requirement for the number of prey fragments (and thus counts) associated with a bait, baits with fewer are skipped.}
}
\value{
List containing the model output directory and the models themselves.

Details.
The steepness of the function to be fitted to putative peaks is determined by \eqn{\omega} according to \eqn{\beta \exp{- \abs{\omega * d}}}, where \beta represents peak height and \eqn{d} the distance from the center of the peak in bp.
}
\description{
Fits additive models of peaks of varying strengths in various locations to the adjusted readcounts via RJMCMC, and stores these models on disk.
}
\examples{
base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

fits_dir = paste0(base,"/fits")

for(bin_index in 1:5){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir,subsample_size=1000)
}

baits_dir = paste0(base,"/baits")

split_baits_fs(bins_dir,residuals_dir = fits_dir, indices=1:5, output_dir = baits_dir)

baitlist = paste0(baits_dir,"/baitlist.txt")
rjmcmc_dir = paste0(base,"/rjmcmc")
omega_power = 4.7
\donttest{
for(i in 1:3){
 peaky_fs(baitlist,i,output_dir=rjmcmc_dir,omega_power=omega_power)
}
}

}
