% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{rjmcmc_list}
\alias{rjmcmc_list}
\title{Create a list of peaky model files}
\usage{
rjmcmc_list(rjmcmc_dir, filename = NULL)
}
\arguments{
\item{filename}{Path}

\item{interactions_file}{Data table containing putative interactions. Columns called baitID, preyID, N, storing the bait fragment ID, prey fragment ID and readcount, respectively.}

\item{bins}{Number of bins to place the interactions into.}

\item{log_file}{Path to a log file.}
}
\value{
List containing the binned interactions ($interactions) and an overview of bins ($bins).
}
\description{
Reports and lists files that contain peaky models in preparation for their interpretation with interpret_peaky_fs()
}
\examples{
base = system.file("extdata",package="peaky")

interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

\donttest{BI = bin_interactions(interactions, fragments, bins=5)}
print(BI)

}
