% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{split_baits}
\alias{split_baits}
\title{Regroups putative interactions by bait}
\usage{
split_baits(bins_dir, residuals_dir, indices, output_dir, plots = TRUE,
  log_file = TRUE)
}
\arguments{
\item{bins_dir}{Directory containing putative interactions that are binned by distance.}

\item{residuals_dir}{Directory where the adjusted read counts from each distance bin are stored.}

\item{indices}{The array of distance bins of which the baits to be processed. These must all have had null models fitted.}

\item{output_dir}{Directory where all putative interactive will be stored, one file per bait. Will be created if it does not exist.}

\item{plots}{Whether adjusted readcounts are to be plotted aganst distance and stored for each bait.}
}
\value{
The output directory.
}
\description{
Regroups putative interactions by bait, rather than by distance, to prepare them for parallel RJMCMC processing. Generates a separate file for each bait. Paths are stored in baitlist.txt, which serves as a to-do list for peaky().
}
\examples{
base = "/local/data/public/hic/demo"
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

\dontrun{bin_interactions(interactions_file, fragments_file, output_dir=bins_dir)}

}
