% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{split_baits_fs}
\alias{split_baits_fs}
\title{Wrapper for split_baits() that interacts with the filesystem}
\usage{
split_baits_fs(bins_dir, residuals_dir, indices, output_dir, plots = TRUE)
}
\arguments{
\item{bins_dir}{Directory containing putative interactions that are binned by distance.}

\item{residuals_dir}{Directory where the adjusted read counts from each distance bin are stored.}

\item{indices}{Indices of distance bins whose baits are processed. These must all have had null models fitted.}

\item{output_dir}{Directory where all putative interactive will be stored, one file per bait. Will be created if it does not exist.}

\item{plots}{Whether adjusted readcounts are to be plotted aganst distance and stored for each bait.}
}
\value{
The output directory.
}
\description{
Calculates p-values under the negative binomial model. Subsequently regroups putative interactions by bait, rather than by distance, to prepare them for parallel RJMCMC processing. Generates a separate file for each bait. Paths are stored in baitlist.txt, which serves as a to-do list for peaky().
}
\examples{
base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

fits_dir = paste0(base,"/fits")

for(bin_index in 1:5){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir,subsample_size=1000)
}

baits_dir = paste0(base,"/baits")

split_baits_fs(bins_dir,residuals_dir = fits_dir, indices=1:5, output_dir = baits_dir)

}
