% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{bin_interactions}
\alias{bin_interactions}
\title{Bins putative interactions by distance}
\usage{
bin_interactions(interactions, fragments, bins = 5, min_dist = 2500,
  max_dist = Inf, log_file = NA)
}
\arguments{
\item{interactions}{Data table containing putative interactions. Columns called baitID, preyID, N, storing the bait fragment ID, prey fragment ID and readcount, respectively.}

\item{fragments}{Data table containing fragment information. Columns called chrom, chromStart, chromEnd, ID, storing the chromosome, starting coordinate, ending coordinate and ID of a fragment, respectively.}

\item{bins}{Number of bins to place the interactions into.}

\item{min_dist}{Minimum distance interactions must span to be included in the analysis. Distance is defined between fragment midpoints.}

\item{max_dist}{Maximum distance interactions can span to be included in the analysis. Distance is defined between fragment midpoints.}

\item{log_file}{Path to a log file.}
}
\value{
List containing the binned interactions ($interactions) and an overview of bins ($bins).
}
\description{
Places putative interactions into equally-sized bins based on the distances they span.
}
\examples{
base = system.file("extdata",package="peaky")

interactions_file = paste0(base,"/counts.tsv")
fragments_file = paste0(base,"/fragments.bed")

interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

\donttest{BI = bin_interactions(interactions, fragments, bins=5)}
print(BI)

}
