% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{bin_interactions}
\alias{bin_interactions}
\title{Bins putative interactions by distance}
\usage{
bin_interactions(interactions_file, fragments_file, output_dir, bins = 50)
}
\arguments{
\item{interactions_file}{Path to (tsv or csv) file containing putative interactions. Its first line must state the columns names: baitID, preyID, N. Subsequent lines report one putative interaction each: a bait fragment ID, prey fragment ID and readcount.}

\item{fragments_file}{Path to (bed) file containing fragment information. Its first line must state the column names: chrom, chromStart, chromEnd, ID. Each subsequent line reports the chromosome, starting coordinate, ending coordinate and ID of a fragment.}

\item{output_dir}{Directory where the generated bins will be stored. Will be created if it does not exist.}

\item{bins}{Number of bins to place the interactions into.}
}
\value{
List containing the output directory and an overview of bins.
}
\description{
Places putative interactions into equally-sized bins based on the distances they span, and stores these on disk.
}
\examples{
base = "/local/data/public/hic/demo"
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

\dontrun{bin_interactions(interactions_file, fragments_file, output_dir=bins_dir)}

}
