% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{peaky}
\alias{peaky}
\title{Obtains peak locations and heights based on adjusted readcounts}
\usage{
peaky(bait, omega_power, iterations = 1e+06, min_interactions = 20,
  log_file = NA)
}
\arguments{
\item{bait}{Data table containing the putative interactions of a bait, having the columns 'baitID', 'dist', and 'residual'. These report the bait ID, its distance to putative preys, and the adjusted readcounts for its interactions with them, respectively.}

\item{omega_power}{Expected decay of adjusted read counts around a truly interacting prey. See details.}

\item{iterations}{Number of models to parametrize. Greated numbers should lead to increased reproducibility.}

\item{log_file}{Path to a log file.}
}
\value{
The output directory.

Details.
The steepness of the function to be fitted to putative peaks is determined by \eqn{\omega} according to \eqn{\beta \exp{- \abs{\omega * d}}}, where \beta represents peak height and \eqn{d} the distance from the center of the peak in bp.
}
\description{
Fits additive models of peaks of varying strengths in various locations to the adjusted readcounts via RJMCMC, and stores these models on disk.
}
\examples{
base = system.file("extdata",package="peaky")

interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

BI = bin_interactions(interactions, fragments, bins=5)

data.table::setorder(BI$interactions, dist.bin)

relevant_bins = 1:8
BI$interactions = BI$interactions[dist.bin\%in\%relevant_bins,]
RES = sapply(relevant_bins, function(x,BI){ model_bin(BI$interactions[dist.bin==x,])$residuals }, BI=BI)

BTS = split_baits(BI$interactions, RES)

peaky(BTS[baitID==618421], omega_power=4.7, iterations=10e3)
}
