% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaky.R
\name{rjmcmc_list}
\alias{rjmcmc_list}
\title{Create a list of peaky model files}
\usage{
rjmcmc_list(rjmcmc_dir, filename = NULL)
}
\arguments{
\item{Path}{to a folder containing the results produced by peaky_fs()}
}
\value{
List containing the path to a report of all results found ($filename), and the paths to the results themselves ($rjmcmc_paths).
}
\description{
Reports and lists files that contain peaky models in preparation for their interpretation with interpret_peaky_fs()
}
